requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
pytest==8.3.3
sqlalchemy==2.0.36
asyncpg==0.29.0
pgvector==0.2.5
httpx[http2]==0.27.2
PyYAML==6.0.2
regex==2024.7.24
structlog==24.4.0
python-dotenv==1.0.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
websockets==12.0
aiohttp==3.10.10
sounddevice==0.4.7
soundfile==0.12.1
numpy==1.26.4
tenacity==8.5.0
pvcheetah==3.0.0
vosk==0.3.44
langdetect==1.0.9
psutil==5.9.8
greenlet==3.0.3
